Move DelayedProcessor inside GroupedCallReducer, where it is used.
